added see more pane, refactored into algolia v2 class and added feature flag
