+++ packages/ui/docs-bundle/src/pages/api/fern-docs/feature-flags.ts
@@ -33,6 +33,7 @@ const FEATURE_FLAGS = [
     "cohere-theme" as const,
     "file-forge-hack-enabled" as const,
     "hide-404-page" as const,
+    "new-search-experience" as const,
 ];
 
 type FeatureFlag = (typeof FEATURE_FLAGS)[number];
@@ -90,6 +91,7 @@ export async function getFeatureFlags(domain: string): Promise<FeatureFlags> {
         const isCohereTheme = checkDomainMatchesCustomers(domain, config["cohere-theme"]);
         const isFileForgeHackEnabled = checkDomainMatchesCustomers(domain, config["file-forge-hack-enabled"]);
         const is404PageHidden = checkDomainMatchesCustomers(domain, config["hide-404-page"]);
+        const isNewSearchExperienceEnabled = checkDomainMatchesCustomers(domain, config["new-search-experience"]);
 
         return {
             isApiPlaygroundEnabled: isApiPlaygroundEnabledOverrides(domain) || isApiPlaygroundEnabled,
@@ -119,6 +121,7 @@ export async function getFeatureFlags(domain: string): Promise<FeatureFlags> {
             isCohereTheme,
             isFileForgeHackEnabled,
             is404PageHidden,
+            isNewSearchExperienceEnabled,
         };
     } catch (e) {
         // eslint-disable-next-line no-console
@@ -151,6 +154,7 @@ export async function getFeatureFlags(domain: string): Promise<FeatureFlags> {
             isCohereTheme: false,
             isFileForgeHackEnabled: false,
             is404PageHidden: false,
+            isNewSearchExperienceEnabled: false,
         };
     }
 }
